% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{calc_catch22}
\alias{calc_catch22}
\title{Calculate catch22 features on a dataframe}
\usage{
calc_catch22(data, catch24)
}
\arguments{
\item{data}{\code{data.frame} containing time-series data}

\item{catch24}{\code{Boolean} specifying whether to compute \code{catch24} in addition to \code{catch22} if \code{catch22} is one of the feature sets selected. Defaults to \code{FALSE}}
}
\value{
\code{data.frame} of features
}
\description{
Calculate catch22 features on a dataframe
}
\author{
Trent Henderson
}
