% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{calc_kats}
\alias{calc_kats}
\title{Calculate Kats features on a dataframe}
\usage{
calc_kats(data)
}
\arguments{
\item{data}{\code{data.frame} containing time-series data}
}
\value{
\code{data.frame} of features
}
\description{
Calculate Kats features on a dataframe
}
\author{
Trent Henderson
}
