% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{calc_tsfresh}
\alias{calc_tsfresh}
\title{Calculate tsfresh features on a dataframe}
\usage{
calc_tsfresh(data, column_id = "id", column_sort = "timepoint", cleanup)
}
\arguments{
\item{data}{\code{data.frame} containing time-series data}

\item{column_id}{\code{character} denoting the id column name. Defaults to \code{"id"}}

\item{column_sort}{\code{character} denoting the timepoint column name. Defaults to \code{"timepoint"}}

\item{cleanup}{\code{character} specifying whether to use the in-built \code{tsfresh} relevant feature filter or not. Defaults to \code{"No"}}
}
\value{
\code{data.frame} of features
}
\description{
Calculate tsfresh features on a dataframe
}
\author{
Trent Henderson
}
