% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_libs.R
\name{install_python_libs}
\alias{install_python_libs}
\title{Download and install all the relevant Python packages into a target location}
\usage{
install_python_libs(python_path, path)
}
\arguments{
\item{python_path}{\code{character} specifying the filepath to the location of Python 3.9 on your machine}

\item{path}{\code{character} denoting the filepath to install the Python libraries and virtual environment to}
}
\description{
Download and install all the relevant Python packages into a target location
}
\author{
Trent Henderson
}
