% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purloiner.R
\docType{package}
\name{purloiner}
\alias{purloiner}
\alias{purloiner-package}
\title{Calculate Time-Series Features from Feature Sets in R and Python}
\description{
Calculate Time-Series Features from Feature Sets in R and Python
}
