% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{tsfeatures_helper}
\alias{tsfeatures_helper}
\title{Helper function to calculate tsfeatures features on a dataframe}
\usage{
tsfeatures_helper(data, grouped = FALSE, feats)
}
\arguments{
\item{data}{\code{data.frame} containing time-series data}
}
\value{
\code{data.frame} of features
}
\description{
Helper function to calculate tsfeatures features on a dataframe
}
\author{
Trent Henderson
}
