% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{calc_catch22}
\alias{calc_catch22}
\title{Calculate catch22 features on a dataframe}
\usage{
calc_catch22(data, catch24)
}
\arguments{
\item{data}{\code{data.frame} containing time-series data}
}
\value{
\code{data.frame} of features
}
\description{
Calculate catch22 features on a dataframe
}
\author{
Trent Henderson
}
