% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{calc_tsfresh}
\alias{calc_tsfresh}
\title{Calculate tsfresh features on a dataframe}
\usage{
calc_tsfresh(data, column_id = "id", column_sort = "timepoint", cleanup)
}
\arguments{
\item{data}{\code{data.frame} containing time-series data}
}
\value{
\code{data.frame} of features
}
\description{
Calculate tsfresh features on a dataframe
}
\author{
Trent Henderson
}
